[PY-WEB-19]BE: Refactor project structure
