+++ lab4/Khalabarchuk_Project/app/forms/ToDo.py
@@ -1,0 +1,8 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField
+from wtforms.validators import DataRequired
+
+
+class ToDoForm(FlaskForm):
+    title = StringField("", validators=[DataRequired(message="This field is required.")])
+    submit = SubmitField("Save")
