+++ lab4/Khalabarchuk_Project/app/templates/comments-page.html
@@ -43,5 +43,4 @@
 
         </main>
     </div>
-
 {% endblock %}
