+++ lab4/Khalabarchuk_Project/app/views.py
@@ -1,15 +1,10 @@
-import datetime
 import platform
 
-from flask import render_template, request, session, redirect, url_for, make_response, flash
-from flask_login import login_user, login_required, current_user, logout_user
+from flask import render_template, request
+from flask_login import login_required
 
 from app import app
-from app import db
-from entitys.Login import LoginForm, ChangePasswordForm
-from entitys.ToDo import ToDo, ToDoForm
-from entitys.Comment import CommentForm, Comment
-from entitys.User import RegistrationForm, AuthUser
+from app.forms.Login import ChangePasswordForm
 
 PROJECTS_LIST = [{
     "photo_url": "projectNMWS.png",
@@ -73,200 +68,11 @@ def base_render(template: str, **context):
     return render_template(template, about_os=platform.platform(), user_agent_info=request.user_agent.string, **context)
 
 
-@app.route("/users")
-@login_required
-def get_users():
-    return base_render("all-users.html", users=AuthUser.query.all())
-
-
-@app.route("/comments")
-@login_required
-def comments_page():
-    return base_render("comments-page.html", comment_form=CommentForm(), comments=Comment.query.all())
-
-
-@app.route("/comments", methods=["POST"])
-@login_required
-def add_comment():
-    comment_form = CommentForm()
-
-    if comment_form.validate_on_submit():
-        date = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-
-        new_comment = Comment(username=current_user.username, comment=comment_form.comment.data, date=date)
-        db.session.add(new_comment)
-        db.session.commit()
-
-        flash('Comment added successfully', 'success')
-    else:
-        flash('Something went wrong while adding a comment', 'danger')
-
-    return redirect(url_for("comments_page"))
-
-
-@app.route('/todo', methods=["GET"])
-@login_required
-def todo_page():
-    return base_render("todo-page.html", todo_list=ToDo.query.all(), todo_form=ToDoForm())
-
-
-@app.route("/todo", methods=["POST"])
-@login_required
-def add_todo():
-    todo_form = ToDoForm()
-    if todo_form.validate_on_submit():
-        new_todo = ToDo(title=todo_form.title.data, completed=False, status=ToDo.Status.IN_PROGRESS)
-        db.session.add(new_todo)
-        db.session.commit()
-
-        flash('Todo added successfully', 'success')
-
-    return redirect(url_for("todo_page"))
-
-
-@app.route("/todo/<string:id>")
-@login_required
-def delete_todo(id: str):
-    todo = db.get_or_404(ToDo, id)
-    db.session.delete(todo)
-    db.session.commit()
-    flash('Todo deleted successfully', 'success')
-
-    return redirect(url_for("todo_page"))
-
-
-@app.route("/todo/<string:id>/update")
-@login_required
-def update_todo(id: str):
-    todo = db.get_or_404(ToDo, id)
-    todo.complete()
-
-    db.session.commit()
-    flash('Todo updated successfully', 'success')
-
-    return redirect(url_for("todo_page"))
-
-
-@app.route("/register", methods=["GET", "POST"])
-def sign_up():
-    if current_user.is_authenticated:
-        return redirect(url_for("info"))
-
-    reg_form = RegistrationForm()
-
-    if reg_form.validate_on_submit():
-        user = AuthUser(email=reg_form.email.data, username=reg_form.username.data,
-                        password=reg_form.password.data)
-        db.session.add(user)
-        db.session.commit()
-
-        flash('You have successfully registered', 'success')
-        return redirect(url_for("login"))
-
-    return base_render("sign_up.html", reg_form=reg_form)
-
-
-@app.route("/account")
-@login_required
-def account():
-    return base_render("account.html")
-
-
-@app.route('/login', methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("info"))
-
-    form = LoginForm()
-
-    if form.validate_on_submit():
-        db_user = db.session.query(AuthUser).filter_by(email=form.login.data).first()
-
-        if not db_user or not db_user.verify_password(password=form.password.data):
-            form.login.errors.append("Login or password is incorrect")
-            form.password.errors.append("Login or password is incorrect")
-            flash("Login or password is incorrect", 'danger')
-            return base_render("login.html", form=form)
-
-        login_user(db_user, remember=form.remember.data)
-
-        flash("You have successfully signed-in", "success")
-        return redirect(url_for("account"))
-
-    return base_render("login.html", form=form)
-
-
-@app.route("/logout")
-@login_required
-def logout():
-    session.clear()
-    logout_user()
-    flash("You have successfully quit", "success")
-    return redirect(url_for("login"))
-
-
-@app.route("/change_password", methods=["POST"])
-@login_required
-def change_password():
-    form = ChangePasswordForm()
-
-    if form.validate_on_submit():
-        db_user = db.session.query(AuthUser).filter_by(email=current_user.email).first()
-
-        if db_user and db_user.verify_password(form.old_password.data):
-            db_user.password = form.new_password.data
-            db.session.add(db_user)
-            db.session.commit()
-            flash("You have successfully changed your password", "success")
-        else:
-            form.old_password.errors.append("Enter the correct old password")
-
-    return base_render("info.html", change_password_from=form, cookies=request.cookies.items())
-
-
 @app.route("/info")
 @login_required
 def info():
     form = ChangePasswordForm()
     return base_render("info.html", change_password_from=form, cookies=request.cookies.items())
-
-
-@app.route("/add_cookie", methods=["POST"])
-@login_required
-def add_cookie():
-    cookie_key = request.form.get("cookie_key")
-    cookie_value = request.form.get("cookie_value")
-    cookie_exp = request.form.get("expires_at")
-
-    response = make_response(redirect(url_for("info")))
-    response.set_cookie(cookie_key, cookie_value, expires=cookie_exp)
-
-    flash("You added a fancy cookie", "info")
-    return response
-
-
-@app.route("/remove_cookie", methods=["POST"])
-@login_required
-def remove_cookie():
-    cookie_key = request.form.get("cookie_key")
-
-    response = make_response(redirect(url_for("info")))
-
-    if cookie_key is None:
-        for key in request.cookies.keys():
-            if key != "session":
-                response.delete_cookie(key)
-
-        flash("You have given all the cookies", "info")
-        return response
-
-    if cookie_key not in request.cookies.keys():
-        flash(f"Undefined {cookie_key}", "danger")
-        return redirect(url_for('info'))
-
-    response.delete_cookie(cookie_key)
-    flash(f"Unfortunately, we have successfully deleted your cookie {cookie_key}", "warning")
-    return response
 
 
 @app.route('/')
