--- lab4/Khalabarchuk_Project/entitys/Registration.py
@@ -1,9 +1,0 @@
-from app import db
-
-
-# class Registration(db.Model):
-#     id = db.Column(db.Integer, primary_key=True, unique=False, nullable=False)
-#     email = db.Column(db.String(35), unique=False, nullable=False)
-#     username = db.Column(db.String(30), unique=False, nullable=False)
-#     password = db.Column(db.String(60), unique=False, nullable=False)
-
