+++ lab4/Khalabarchuk_Project/entitys/ToDo.py
@@ -1,13 +1,4 @@
-from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField
-from wtforms.validators import DataRequired
-
 from app import db
-
-
-class ToDoForm(FlaskForm):
-    title = StringField("", validators=[DataRequired(message="This field is required.")])
-    submit = SubmitField("Save")
 
 
 class ToDo(db.Model):
